requirements.txt
fastapi==0.111.1
uvicorn
sqlalchemy==2.0.22
pyodbc
python-dotenv==1.0.1
google-generativeai
pydantic